+++ OUDS/Core/Themes/Orange/Sources/Providers/ComponentTokens/OrangeThemeBulletListComponentTokensProvider.swift
@@ -113,8 +113,8 @@ open class OrangeThemeBulletListComponentTokensProvider: AllBulletListComponentT
 
     // ଘ( ･ω･)_/ﾟ･:*:･｡☆
     // Note: So as to help the integration of generated code produced by the tokenator
-    // the implemention of BreadcrumbComponentTokens is not here but in Core/Themes/Orange/Values/ComponentTokens/OrangeTheme+BreadcrumbComponentTokens.swift
-    // This declaration of OrangeThemeBreadcrumbComponentTokensProvider is here also to allow to write documentation.
+    // the implemention of BulletListComponentTokens is not here but in Core/Themes/Orange/Values/ComponentTokens/OrangeTheme+BulletListComponentTokens.swift
+    // This declaration of OrangeThemeBulletListComponentTokensProvider is here also to allow to write documentation.
 }
 
 // swiftlint:enable type_name